refactor(auth): move auth datasource to external module and update imports
Update all imports of AuthDatasource to point to the new location in the external auth module
Delete old test files and implement new auth datasource with all functionality
