+++ src/infrastructure/external/auth/datasources/__tests__/logout.test.ts
@@ -1,4 +1,4 @@
-import { AuthDatasource } from "@/infrastructure/datasources/auth.datasource";
+import { AuthDatasource } from "@/infrastructure/external/auth/datasources/auth.datasource";
 import { AuthClient } from "@/infrastructure/external/auth/auth.client";
 import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { createMockLogoutDto } from "@/config/__tests__/__helpers__/auth-datasource.helpers";
